Refactor socket controller logic
