+++ matching-service/src/routes/socket-routes.ts
@@ -1,0 +1,8 @@
+import { Server, Socket } from "socket.io";
+import { handleDisconnect, handleRegisterForMatching } from "../controller/socket-controller";
+
+export function registerEventHandlers(socket: Socket, io: Server) {
+    handleRegisterForMatching(socket, io);
+    handleDisconnect(socket);
+}
+
